🚚 [app] => move `local_auth.py` in `environment`
