+++ src/youwol/app/middlewares/__init__.py
@@ -1,5 +1,4 @@
 # relative
 from .browser_caching_middleware import *
 from .hybridizer_middleware import *
-from .local_auth import *
 from .local_cloud_hybridizers import *
